Praticando Vetores e Matrizes

+++ src/JavaBasico/Ex2.java
@@ -10,15 +10,15 @@ public class Ex2 {
     public static void main(String[] args) {
         Scanner entrada = new Scanner(System.in);
 
-        double nota1,nota2,nota3,nota4,media;
+        float nota1,nota2,nota3,nota4,media;
         System.out.println("Digite sua primeira nota: ");
-        nota1 = entrada.nextDouble();
+        nota1 = entrada.nextFloat();
         System.out.println("Digite sua segunda nota: ");
-        nota2 = entrada.nextDouble();
+        nota2 = entrada.nextFloat();
         System.out.println("Digite sua terceira nota: ");
-        nota3 = entrada.nextDouble();
+        nota3 = entrada.nextFloat();
         System.out.println("Digite sua quarta nota: ");
-        nota4 = entrada.nextDouble();
+        nota4 = entrada.nextFloat();
 
         media = (nota1+nota2+nota3+nota4) / 4;
